requirements.txt
# Core ML and AI Libraries
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
transformers==4.56.2
diffusers @ git+https://github.com/huggingface/diffusers.git@1448b035859dd57bbb565239dcdd79a025a85422
accelerate==1.10.1
safetensors==0.6.2
tokenizers==0.22.1
sentencepiece==0.2.1

# DiffSynth Studio
-e git+https://github.com/modelscope/DiffSynth-Studio.git@ed256ef8be195d5deae2846a7e9f025670d99db3#egg=diffsynth

# ModelScope and HuggingFace
modelscope==1.30.0
huggingface-hub==0.35.0

# FastAPI and Web Framework
fastapi==0.117.1
uvicorn==0.36.0
starlette==0.48.0
pydantic==2.11.9
pydantic_core==2.33.2
python-multipart==0.0.20
aiofiles==24.1.0

# Image Processing
pillow==11.3.0
opencv-python==4.12.0.88
imageio==2.37.0
imageio-ffmpeg==0.6.0

# Scientific Computing
numpy==2.2.6
scipy==1.16.2
pandas==2.3.2
einops==0.8.1

# NVIDIA CUDA Support
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90

# Memory and Performance Optimization
flash_attn==2.8.3
xformers==0.0.32.post2
bitsandbytes==0.47.0
triton==3.4.0
peft==0.17.1

# Gradio UI (Optional)
gradio==5.46.1
gradio_client==1.13.1

# Utilities
requests==2.32.5
tqdm==4.67.1
regex==2025.9.18
ftfy==6.3.1
packaging==25.0
filelock==3.19.1
fsspec==2025.9.0
psutil==7.1.0
python-dotenv==1.1.1

# Development and Testing
pytest==8.4.2
black==25.9.0
flake8==7.3.0
mypy==1.18.2
isort==6.0.1
ruff==0.13.1
pre_commit==4.3.0
bandit==1.8.6
codespell==2.4.1

# Audio Processing (for Gradio)
pydub==0.25.1
ffmpy==0.6.1

# Web and HTTP
httpx==0.28.1
httpcore==1.0.9
httptools==0.6.4
websockets==15.0.1
uvloop==0.21.0
watchfiles==1.1.0

# Configuration and Serialization
PyYAML==6.0.2
tomlkit==0.13.3
protobuf==6.32.1

# CLI and Terminal
typer==0.19.1
rich==14.1.0
click==8.3.0

# Other Dependencies
Jinja2==3.1.6
MarkupSafe==3.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
six==1.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2